fix delete pics issu

+++ handlers/fileSystem.js
@@ -2,9 +2,14 @@ const path = require("path");
 const fs = require("fs");
 const directory = "public/uploads";
 
-exports.deleteAllImagesInUploadsFolder = (req, res, next) => {
+const goOnAndDelete = () => {
   fs.readdir(directory, (err, files) => {
     if (err) throw err;
+    
+    if (files.length === 0) {
+      console.log("Folder is empty!");
+      return;
+    }
 
     for (const file of files) {
       fs.unlink(path.join(directory, file), err => {
@@ -13,3 +18,37 @@ exports.deleteAllImagesInUploadsFolder = (req, res, next) => {
     }
   });
 }
+
+exports.deleteAllImagesInUploadsFolder = () => {
+  //check if folder exists 
+  fs.access(directory, error => {
+    if (error) {
+      console.log("Directory does not exist.");
+      return;
+    } else {
+      console.log("Directory exists.");
+      goOnAndDelete();
+    }
+  });
+}
+
+
+// exports.deleteAllImagesInUploadsFolder = (req, res, next) => {
+//   try {
+//     fs.readdir(directory, (err, files) => {
+//       if (err) throw err;
+
+//       for (const file of files) {
+//         fs.unlink(path.join(directory, file), err => {
+//           if (err) throw err;
+//         });
+//       }
+//     });
+
+//     res
+//       .status(200)
+//       .json({ message: "All images in uploads folder cleared up" });
+//   } catch (err) {
+//     return next(err);
+//   }
+// };
